bump to v27.0.0-rc.7

+++ include/autd3/driver/datagram/stm/gain.hpp
@@ -34,41 +34,41 @@ class GainSTM final : public IntoDatagramTuple<GainSTM>,
       : _loop_behavior(LoopBehavior::Infinite), _mode(native_methods::GainSTMMode::PhaseIntensityFull) {
     for (auto e : iter)
       _gains.emplace_back(std::make_shared<std::remove_reference_t<std::ranges::range_value_t<R>>>(std::forward<std::ranges::range_value_t<R>>(e)));
-    _sampling = STMSamplingConfig(sampling, static_cast<uint32_t>(_gains.size()));
+    _sampling = STMSamplingConfig(sampling, static_cast<uint16_t>(_gains.size()));
   }
   template <stm_sampling_config T, gain G>
   AUTD3_API explicit GainSTM(const T sampling, std::initializer_list<G> iter)
       : _loop_behavior(LoopBehavior::Infinite), _mode(native_methods::GainSTMMode::PhaseIntensityFull) {
     for (auto e : iter) _gains.emplace_back(std::make_shared<G>(e));
-    _sampling = STMSamplingConfig(sampling, static_cast<uint32_t>(_gains.size()));
+    _sampling = STMSamplingConfig(sampling, static_cast<uint16_t>(_gains.size()));
   }
 
   template <gain_range R>
   AUTD3_API [[nodiscard]] static GainSTM nearest(const Freq<float> freq, const R& iter) {
     std::vector<std::shared_ptr<GainBase>> gains;
     for (auto e : iter) gains.emplace_back(std::make_shared<std::remove_reference_t<std::ranges::range_value_t<R>>>(e));
-    const auto n = static_cast<uint32_t>(gains.size());
+    const auto n = static_cast<uint16_t>(gains.size());
     return GainSTM(STMSamplingConfig::nearest(freq, n), std::move(gains));
   }
   template <gain G>
   AUTD3_API [[nodiscard]] static GainSTM nearest(const Freq<float> freq, std::initializer_list<G> iter) {
     std::vector<std::shared_ptr<GainBase>> gains;
     for (auto e : iter) gains.emplace_back(std::make_shared<G>(e));
-    const auto n = static_cast<uint32_t>(gains.size());
+    const auto n = static_cast<uint16_t>(gains.size());
     return GainSTM(STMSamplingConfig::nearest(freq, n), std::move(gains));
   }
   template <typename Rep, typename P, gain_range R>
   AUTD3_API [[nodiscard]] static GainSTM nearest(const std::chrono::duration<Rep, P> period, const R& iter) {
     std::vector<std::shared_ptr<GainBase>> gains;
     for (auto e : iter) gains.emplace_back(std::make_shared<std::remove_reference_t<std::ranges::range_value_t<R>>>(e));
-    const auto n = static_cast<uint32_t>(gains.size());
+    const auto n = static_cast<uint16_t>(gains.size());
     return GainSTM(STMSamplingConfig::nearest(period, n), std::move(gains));
   }
   template <typename Rep, typename P, gain G>
   AUTD3_API [[nodiscard]] static GainSTM nearest(const std::chrono::duration<Rep, P> period, std::initializer_list<G> iter) {
     std::vector<std::shared_ptr<GainBase>> gains;
     for (auto e : iter) gains.emplace_back(std::make_shared<G>(e));
-    const auto n = static_cast<uint32_t>(gains.size());
+    const auto n = static_cast<uint16_t>(gains.size());
     return GainSTM(STMSamplingConfig::nearest(period, n), std::move(gains));
   }
